feat: implementasi sistem poin - min belanja Rp10.000 dapat 10 poin, redeem 100 poin = diskon Rp1.000

+++ lib/ui/customer/home/home_screen.dart
@@ -198,7 +198,7 @@ class HomeScreen extends StatelessWidget {
   // Header Section
   // ═══════════════════════════════════════════════════════════════════
 
-  Widget _buildHeader(BuildContext context, String firstName, int poin) {
+  Widget _buildHeader(BuildContext context, String firstName, double poin) {
     return Container(
       padding: const EdgeInsets.fromLTRB(24, 16, 24, 28),
       decoration: const BoxDecoration(
